[COLAB-2340] moved deleteOrphanProposals
